Enable S-Curve Acceleration and tweak jerk
To further optimize print quality even at very high speeds (I aim for roughly 80-100mm/s), I enabled S_CURVE_ACCELERATION and tweaked the jerk and default acceleration values.

Note: This is still in testing. As soon as I finish and it works as I like, I'll push it onto the master branch. If you see this commit on the master branch, it's safe to flash.